Refacto code into headers and sverals files
