+++ .vscode/settings.json
@@ -1,0 +1,12 @@
+{
+    "files.associations": {
+        "string": "cpp",
+        "array": "cpp",
+        "deque": "cpp",
+        "unordered_map": "cpp",
+        "unordered_set": "cpp",
+        "vector": "cpp",
+        "initializer_list": "cpp",
+        "regex": "cpp"
+    }
+}
