+++ src/led.cpp
@@ -1,0 +1,43 @@
+#include "FastLED.h"
+#include "constants.h"
+
+void setLedColor()
+{
+  for (int i = 0; i < NUM_LEDS; i++)
+  {
+    leds[i] = CRGB::Black;
+  }
+  FastLED.show();
+}
+
+void flashLed()
+{
+  for (int i = 0; i < NUM_LEDS; i++)
+  {
+    leds[i] = CRGB::Blue;
+  }
+  FastLED.show();
+  for (int i = 0; i < NUM_LEDS; i++)
+  {
+    leds[i] = CRGB::Black;
+  }
+  FastLED.show();
+}
+
+void CircleColor(uint8_t R, uint8_t G, uint8_t B)
+{
+  for (int i = 0; i <= NUM_LEDS; i++)
+  {
+    leds[i] = CRGB(R, G, B);
+    FastLED.show();
+
+    delay(50);
+  }
+}
+
+void initLeds()
+{
+  FastLED.addLeds<WS2812, LED_PIN, GRB>(leds, NUM_LEDS);
+  FastLED.setBrightness(max_bright);
+  set_max_power_in_volts_and_milliamps(5, 1000);
+}
